Change telegram_id type

+++ priv/repo/migrations/20201205205445_add_telegram_fields_to_users.exs
@@ -4,7 +4,7 @@ defmodule Squadster.Repo.Migrations.AddTelegramFieldsToUsers do
   def change do
     alter table(:users) do
       add :telegram_chat_id, :bigint
-      add :telegram_id, :bigint
+      add :telegram_id, :integer
       add :telegram_token, :string
     end
